#10900 cleanup breadcrumbs/menu requests
