+++ frontend/src/actions/MenuActions.js
@@ -1,59 +1,7 @@
-import axios from 'axios';
 import { HOME_MENU_USER_MAX_ITEMS } from '../constants/Constants';
 
 import * as types from '../constants/MenuTypes';
-
-export function pathRequest(nodeId) {
-  return axios.get(
-    config.API_URL + '/menu/' + nodeId + '/path/' + '&inclusive=true'
-  );
-}
-
-export function nodePathsRequest(nodeId, limit) {
-  return axios.get(
-    config.API_URL +
-      '/menu/node/' +
-      nodeId +
-      '?depth=2' +
-      (limit ? '&childrenLimit=' + limit : '')
-  );
-}
-
-export function elementPathRequest(pathType, elementId) {
-  return axios.get(
-    config.API_URL +
-      '/menu/elementPath?type=' +
-      pathType +
-      '&elementId=' +
-      elementId +
-      '&inclusive=true'
-  );
-}
-
-export function queryPathsRequest(query, limit, child) {
-  return axios.get(
-    config.API_URL +
-      '/menu/queryPaths?nameQuery=' +
-      query +
-      (limit ? '&childrenLimit=' + limit : '') +
-      (child ? '&childrenInclusive=true' : '')
-  );
-}
-
-export function rootRequest(limit, depth = 0, onlyFavorites) {
-  return axios.get(
-    config.API_URL +
-      '/menu/root?depth=' +
-      depth +
-      (limit ? '&childrenLimit=' + limit : '') +
-      (onlyFavorites ? '&favorites=true' : '')
-  );
-}
-export function breadcrumbRequest(nodeId) {
-  return axios.get(config.API_URL + '/menu/node/' + nodeId + '/breadcrumbMenu');
-}
-
-// END OF REQUESTS
+import { elementPathRequest, rootRequest } from '../api';
 
 export function setBreadcrumb(breadcrumb) {
   return {
