+++ frontend/src/api/index.js
@@ -54,6 +54,11 @@ import {
 import {
   getNotificationsRequest,
   getNotificationsEndpointRequest,
+  pathRequest,
+  nodePathsRequest,
+  elementPathRequest,
+  queryPathsRequest,
+  rootRequest,
 } from './app';
 
 export {
@@ -65,6 +70,7 @@ export {
   deleteStaticFilter,
   deleteViewRequest,
   discardNewRequest,
+  elementPathRequest,
   filterViewRequest,
   formatParentUrl,
   getAttributesInstance,
@@ -90,7 +96,6 @@ export {
   getZoomIntoWindow,
   headerPropertiesRequest,
   initQuickInput,
-  setUserLang,
   localLoginRequest,
   locationConfigRequest,
   locationSearchRequest,
@@ -98,12 +103,17 @@ export {
   loginRequest,
   loginWithToken,
   logoutRequest,
+  nodePathsRequest,
   patchRequest,
+  pathRequest,
   patchViewAttributesRequest,
+  queryPathsRequest,
   quickActionsRequest,
+  rootRequest,
   resetPasswordComplete,
   resetPasswordGetAvatar,
   resetPasswordRequest,
+  setUserLang,
   startProcess,
   topActionsRequest,
 };
